Create GUI_framework.py
This is based on the api_test.py file which allows connection to the wearable. It adds a GUI which is used to connect/disconnect from the wearable as well as a framework for how we could integrate the 3 value-added so we can access them from the same start page.